moving buildOptionParam to utils and adding a unit test for reducing code redundancy
