--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/database
@@ -1,1 +1,0 @@
-db1
