--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/host
@@ -1,1 +1,0 @@
-127.0.0.1
