--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/options
@@ -1,1 +1,0 @@
---cluster=fruit-service-test&opt2=val2&optinvalid
