--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/password
@@ -1,1 +1,0 @@
-p1
