--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/port
@@ -1,1 +1,0 @@
-1234
