--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/provider
@@ -1,1 +1,0 @@
-CockroachDB Cloud
