--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/type
@@ -1,1 +1,0 @@
-postgresql
