--- test/test-postgresql-bindings/bindings/postgresql-rhoda-opt/username
@@ -1,1 +1,0 @@
-fruit-user
