+++ test/test-postgresql-bindings/mappedDataBindings/index.js
@@ -79,17 +79,6 @@ const postgresqlDBRHODA = {
   provider: 'CockroachDB Cloud'
 };
 
-const postgresqlDBRHODAOPT = {
-  database: 'db1',
-  host: '127.0.0.1',
-  password: 'p1',
-  port: '1234',
-  options: '--cluster=fruit-service-test -c opt2=val2',
-  type: 'postgresql',
-  user: 'fruit-user',
-  provider: 'CockroachDB Cloud'
-};
-
 const crunchyDataBindingsMappedForOdbc = (function () {
   const { 'ca.crt': caCrt, ...rest } = crunchyDB;
   return {
@@ -100,10 +89,6 @@ const crunchyDataBindingsMappedForOdbc = (function () {
 
 const postgresDataBindingsMappedForPGRHODA = {
   ...postgresqlDBRHODA
-};
-
-const postgresDataBindingsMappedForPGRHODAOPT = {
-  ...postgresqlDBRHODAOPT
 };
 
 function getPqopt({ serviceBindingRoot, overrideSslkeyPath }) {
@@ -150,7 +135,6 @@ module.exports = {
   crunchyDataBindingsMappedForPG,
   crunchyDataBindingsMappedForOdbc,
   postgresDataBindingsMappedForPGRHODA,
-  postgresDataBindingsMappedForPGRHODAOPT,
   getPqopt,
   getSslcerts,
   getOdbcConnectionString
